refactor: modify sourceCode type, add meta types to LoadedItem
