+++ packages/archive-loader-vue/src/createLoader.ts
@@ -5,7 +5,8 @@
  * found in the LICENSE file at https://github.com/IDuxFE/archive/blob/main/LICENSE
  */
 
-import type { ArchiveLoaderVue, ArchiveLoaderVueOptions, SourceCode } from './types'
+import type { ArchiveLoaderVue, ArchiveLoaderVueOptions } from './types'
+import type { SourceCode } from '@idux/archive-vite-plugin'
 
 import { readFileSync } from 'node:fs'
