+++ packages/archive-loader-vue/src/genSourceCodes.ts
@@ -5,7 +5,7 @@
  * found in the LICENSE file at https://github.com/IDuxFE/archive/blob/main/LICENSE
  */
 
-import type { SourceCode } from './types'
+import type { SourceCode } from '@idux/archive-vite-plugin'
 
 import { type MarkdownRenderer } from '@idux/archive-vite-markdown-plugin'
